backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
pydantic==2.9.2
httpx==0.27.2
beautifulsoup4==4.12.3
python-dateutil==2.9.0
lxml==5.3.0
python-multipart==0.0.12
python-dotenv==1.0.1
playwright>=1.44.0
